+++ src/core/langs/java.ts
@@ -88,6 +88,13 @@ export class LangJava extends Lang {
                 outputPath,
             });
             Io.compilationMsg = result.stderr.trim();
+            if (ac.signal.aborted) {
+                this.logger.warn('Compilation aborted by user');
+                return {
+                    verdict: TCVerdicts.RJ,
+                    msg: vscode.l10n.t('Compilation aborted by user.'),
+                };
+            }
             return {
                 verdict: await access(outputPath, constants.R_OK)
                     .then(() => TCVerdicts.UKE)
@@ -97,13 +104,6 @@ export class LangJava extends Lang {
             };
         } catch (e) {
             this.logger.error('Compilation failed', e);
-            if (ac.signal.aborted) {
-                this.logger.warn('Compilation aborted by user');
-                return {
-                    verdict: TCVerdicts.RJ,
-                    msg: vscode.l10n.t('Compilation aborted by user.'),
-                };
-            }
             Io.compilationMsg = (e as Error).message;
             return { verdict: TCVerdicts.CE, msg: '' };
         }