+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the "cph-ng" extension will be documented in this file.
 
+## [Unreleased]
+
+- **fix**: ensure document is saved before proceeding with compilation
+
 ## 0.2.2
 
 - **feat**: add default file extension setting for Competitive Companion