+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to the "cph-ng" extension will be documented in this file.
 
 - **fix**: ensure document is saved before proceeding with compilation
 - **fix**: add message for compilation timeout
+- **style**: rename confusing variable names
 
 ## 0.2.2