+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to the "cph-ng" extension will be documented in this file.
 - **feat**: clear test case result
 - **feat**: add custom file system
 - **feat**: edit raw problem data
+- **fix**: show reject when cancelling compile
 
 ## 0.2.2