+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the "cph-ng" extension will be documented in this file.
 
-## [Unreleased]
+## 0.2.3
 
 - **fix**: ensure document is saved before proceeding with compilation
 - **fix**: add message for compilation timeout